+++ imagepicker/src/androidTest/java/com/github/arykoz/imagepicker/ExampleInstrumentedTest.kt
@@ -1,0 +1,20 @@
+package com.github.arykoz.imagepicker
+
+import androidx.test.InstrumentationRegistry
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+
+        assertEquals("com.github.arykoz.imagepicker.test", appContext.packageName)
+    }
+}
