+++ imagepicker/src/main/kotlin/com/github/arykoz/imagepicker/ImagePickerFileProvider.kt
@@ -1,0 +1,5 @@
+package com.github.arykoz.imagepicker
+
+import androidx.core.content.FileProvider
+
+class ImagePickerFileProvider : FileProvider()
