+++ imagepicker/src/main/kotlin/com/github/arykoz/imagepicker/constant/ImageProvider.kt
@@ -1,0 +1,15 @@
+package com.github.arykoz.imagepicker.constant
+
+/**
+ * Define Image Provider
+ *
+ * @author Dhaval Patel
+ * @version 1.0
+ * @since 04 January 2019
+ */
+enum class ImageProvider {
+    GALLERY,
+    CAMERA,
+    FRONT_CAMERA,
+    BOTH
+}
