--- imagepicker/src/main/kotlin/com/github/dhaval2404/imagepicker/ImagePickerFileProvider.kt
@@ -1,5 +1,0 @@
-package com.github.dhaval2404.imagepicker
-
-import androidx.core.content.FileProvider
-
-class ImagePickerFileProvider : FileProvider()
