--- imagepicker/src/main/kotlin/com/github/dhaval2404/imagepicker/constant/ImageProvider.kt
@@ -1,15 +1,0 @@
-package com.github.dhaval2404.imagepicker.constant
-
-/**
- * Define Image Provider
- *
- * @author Dhaval Patel
- * @version 1.0
- * @since 04 January 2019
- */
-enum class ImageProvider {
-    GALLERY,
-    CAMERA,
-    FRONT_CAMERA,
-    BOTH
-}
