--- imagepicker/src/main/kotlin/com/github/dhaval2404/imagepicker/util/DialogHelper.kt
@@ -1,51 +1,0 @@
-package com.github.dhaval2404.imagepicker.util
-
-import android.content.Context
-import android.view.LayoutInflater
-import androidx.appcompat.app.AlertDialog
-import com.github.dhaval2404.imagepicker.R
-import com.github.dhaval2404.imagepicker.constant.ImageProvider
-import com.github.dhaval2404.imagepicker.listener.ResultListener
-import kotlinx.android.synthetic.main.dialog_choose_app.view.*
-
-/**
- * Show Dialog
- *
- * @author Dhaval Patel
- * @version 1.0
- * @since 04 January 2018
- */
-internal object DialogHelper {
-
-    /**
-     * Show Image Provide Picker Dialog. This will streamline the code to pick/capture image
-     *
-     */
-    fun showChooseAppDialog(context: Context, listener: ResultListener<ImageProvider>) {
-        val layoutInflater = LayoutInflater.from(context)
-        val customView = layoutInflater.inflate(R.layout.dialog_choose_app, null)
-
-        val dialog = AlertDialog.Builder(context)
-            .setTitle(R.string.title_choose_image_provider)
-            .setView(customView)
-            .setOnCancelListener {
-                listener.onResult(null)
-            }
-            .setNegativeButton(R.string.action_cancel) { _, _ ->
-                listener.onResult(null)
-            }
-            .show()
-
-        // Handle Camera option click
-        customView.lytCameraPick.setOnClickListener {
-            listener.onResult(ImageProvider.CAMERA)
-            dialog.dismiss()
-        }
-
-        // Handle Gallery option click
-        customView.lytGalleryPick.setOnClickListener {
-            listener.onResult(ImageProvider.GALLERY)
-            dialog.dismiss()
-        }
-    }
-}
