--- sample/src/main/kotlin/com.github.dhaval2404.imagepicker/sample/ImageViewExtension.kt
@@ -1,25 +1,0 @@
-package com.github.dhaval2404.imagepicker.sample
-
-import android.widget.ImageView
-import androidx.annotation.DrawableRes
-import com.bumptech.glide.Glide
-import com.bumptech.glide.request.RequestOptions
-import java.io.File
-
-fun ImageView.setDrawableImage(@DrawableRes resource: Int, applyCircle: Boolean = false) {
-    val glide = Glide.with(this).load(resource)
-    if (applyCircle) {
-        glide.apply(RequestOptions.circleCropTransform()).into(this)
-    } else {
-        glide.into(this)
-    }
-}
-
-fun ImageView.setLocalImage(file: File, applyCircle: Boolean = false) {
-    val glide = Glide.with(this).load(file)
-    if (applyCircle) {
-        glide.apply(RequestOptions.circleCropTransform()).into(this)
-    } else {
-        glide.into(this)
-    }
-}
